requirements.txt
discord.py==1.7.3
textblob==0.17.1
wiktionaryparser==0.0.97
